calendar.add needs daysValid as Int => conversion when using it

+++ dataland-e2etests/src/test/kotlin/org/dataland/e2etests/utils/DatesHandler.kt
@@ -8,7 +8,7 @@ class DatesHandler {
     private fun addDaysToDate(date: Date, days: Long): Date {
         val calendar = Calendar.getInstance()
         calendar.time = date
-        calendar.add(Calendar.DATE, days)
+        calendar.add(Calendar.DATE, days.toInt())
         return calendar.time
     }
